[CSE-43] refactor: 당첨되지 않았을 경우 원상복구 구현

+++ modules/draw/draw-api/src/main/java/com/eodi/bium/draw/view/DrawPointView.java
@@ -2,7 +2,7 @@ package com.eodi.bium.draw.view;
 
 public record DrawPointView(
     String memberId,
-    Integer point
+    Long point
 ) {
 
 }
